updated homework

+++ src/services/contacts.js
@@ -7,12 +7,21 @@ export const getContacts = async ({
   perPage = 10,
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
-  userId,
+  filter = {},
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactsCollection.find(userId);
+  const contactsQuery = await ContactsCollection.find();
+
+  if(filter.isFavourite) {
+    contactsQuery.where("isFavourite").equals(filter.isFavourite);
+  };
+
+  if(filter.userId) {
+    contactsQuery.where('userId').equals(filter.userId);
+  }
+
   const contactsCount = await ContactsCollection.find()
     .merge(contactsQuery)
     .countDocuments();
@@ -31,15 +40,40 @@ export const getContacts = async ({
   };
 };
 
-export const getContactById = (id) => ContactsCollection.findById(id);
+export const getContactById = ({id, userId}) => {
+  const contact = ContactsCollection.findOne({_id: id, userId});
+  return contact;
+};
 
 export const postContact = async (payload) => {
   const contact = await ContactsCollection.create(payload);
   return contact;
 };
 
-export const deleteContact = async (_id) => {
-  const contact = await ContactsCollection.findOneAndDelete({ _id: _id });
+export const updateContact = async (contactId, payload, userId, options = {}) => {
+
+  const rawResult = await ContactsCollection.findOneAndUpdate(
+    { _id: contactId, userId},
+    payload,
+    {
+      new: true,
+      includeResultMetadata: true,
+      ...options,
+    },
+  );
+
+  if (!rawResult || !rawResult.value) return null;
+
+  return {
+    contact: rawResult.value,
+    isNew: Boolean(rawResult?.lastErrorObject?.upserted),
+  };
+};
+
+
+
+export const deleteContact = async (_id, userId) => {
+  const contact = await ContactsCollection.findOneAndDelete({ _id: _id, userId });
 
   return contact;
 };
